fixed leak when lstnewscmd fails
when it failed only the smcd node was freed and not the cmd struct that;s created in the function. The free_struct function was deconstructed so the free cmd part could be used seperately and put into a new file. to avoid double free's the bool in free tokenlist function after make cmdlist fails in the parser is now set to false. all functions are added to header